Detect max probability of collision, and set the trajectory's fitness value equal to that prob * collision penalty
Former-commit-id: cb37c414e64b56d79a5eeba08ec8bdd520e0e200

+++ trajectory_evaluation/include/collision_detection.h
@@ -30,7 +30,7 @@ class CollisionDetection
     {
       bool inner_;
       bool outer_;
-      double p_max_;
+      int p_max_;
     };
